scripts/requirements.txt
requests>=2.31.0
Pillow>=10.0.0
tqdm>=4.66.0
pykew>=0.1.0
